+++ IncrementalGeneratorExtensions.Content/Datacute/IncrementalGeneratorExtensions/EquatableImmutableArray.cs
@@ -22,112 +22,14 @@ namespace Datacute.IncrementalGeneratorExtensions
     {
         public static EquatableImmutableArray<T> Empty { get; } = new EquatableImmutableArray<T>(ImmutableArray<T>.Empty, 0);
 
-        // The source generation pipelines compare these a lot
-        // so being able to quickly tell when they are different
-        // is important.
-        // We will use an instance cache to find when we can reuse
-        // an existing object, massively speeding up the Equals call.
-        #region Instance Cache
-
-        // The WeakReference allows the GC to collect arrays that are no longer in use.
-        // Thread-safe cache using dictionary of hash code -> list of arrays with that hash
-        private static readonly ConcurrentDictionary<int, List<WeakReference<EquatableImmutableArray<T>>>> ValueCache = new ConcurrentDictionary<int, List<WeakReference<EquatableImmutableArray<T>>>>();
-
         // Static factory method with singleton handling
-        public static EquatableImmutableArray<T> Create(ImmutableArray<T> values, CancellationToken cancellationToken = default)
-        {
-#if !DATACUTE_EXCLUDE_LIGHTWEIGHTTRACE && !DATACUTE_EXCLUDE_GENERATORSTAGE
-            // Record a histogram of the array sizes we are being asked to create
-            LightweightTrace.IncrementCount(GeneratorStage.EquatableImmutableArrayLength, values.Length);
-#endif
-            if (values.IsEmpty)
-                return Empty;
-
-            var hash = CalculateHashCode(values);
-            var list = ValueCache.GetOrAdd(hash, _ => new List<WeakReference<EquatableImmutableArray<T>>>());
-
-            lock (list)
-            {
-                for (int i = list.Count - 1; i >= 0; i--)
-                {
-#if !DATACUTE_EXCLUDE_LIGHTWEIGHTTRACEEXTENSIONS && !DATACUTE_EXCLUDE_GENERATORSTAGE
-                    cancellationToken.ThrowIfCancellationRequested(0);
-#else
-                    cancellationToken.ThrowIfCancellationRequested();
-#endif
-                    if (list[i].TryGetTarget(out var existing))
-                    {
-                        if (ValuesEqual(values, existing._values))
-                        {
-                            // Cache Hit
-#if !DATACUTE_EXCLUDE_LIGHTWEIGHTTRACE && !DATACUTE_EXCLUDE_GENERATORSTAGE
-                            LightweightTrace.IncrementCount(GeneratorStage.EquatableImmutableArrayCacheHit);
-#endif
-                            return existing;
-                        }
-                    }
-                    else
-                    {
-#if !DATACUTE_EXCLUDE_LIGHTWEIGHTTRACE && !DATACUTE_EXCLUDE_GENERATORSTAGE
-                        LightweightTrace.IncrementCount(GeneratorStage.EquatableImmutableArrayCacheWeakReferenceRemoved);
-#endif
-                        list.RemoveAt(i);
-                    }
-                }
-
-                // Cache Miss: Create a new instance
-#if !DATACUTE_EXCLUDE_LIGHTWEIGHTTRACE && !DATACUTE_EXCLUDE_GENERATORSTAGE
-                LightweightTrace.IncrementCount(GeneratorStage.EquatableImmutableArrayCacheMiss);
-#endif
-                var newResult = new EquatableImmutableArray<T>(values, hash);
-                list.Add(new WeakReference<EquatableImmutableArray<T>>(newResult));
-                return newResult;
-            }
-        }
-
-        private static int CalculateHashCode(ImmutableArray<T> values)
-        {
-            var comparer = EqualityComparer<T>.Default;
-            var hash = 0;
-            for (var index = 0; index < values.Length; index++)
-            {
-                var value = values[index];
-                hash = HashHelpers_Combine(hash, value == null ? 0 : comparer.GetHashCode(value));
-            }
-            return hash;
-        }
-        
-        private static int HashHelpers_Combine(int h1, int h2)
-        {
-            // RyuJIT optimizes this to use the ROL instruction
-            // Related GitHub pull request: https://github.com/dotnet/coreclr/pull/1830
-            uint rol5 = ((uint)h1 << 5) | ((uint)h1 >> 27);
-            return ((int)rol5 + h1) ^ h2;
-        }
-
-        private static bool ValuesEqual(ImmutableArray<T> a, ImmutableArray<T> b)
-        {
-            // Identical arrays reference check
-            if (a == b) return true;
-            
-            int length = a.Length;
-            if (length != b.Length) return false;
-            
-            var comparer = EqualityComparer<T>.Default;
-            for (int i = 0; i < length; i++)
-            {
-                if (!comparer.Equals(a[i], b[i]))
-                    return false;
-            }
-            
-            return true;
-        }
-        
-        #endregion
+        public static EquatableImmutableArray<T> Create(ImmutableArray<T> values, CancellationToken cancellationToken = default) 
+            => EquatableImmutableArrayInstanceCache<T>.GetOrCreate(values, cancellationToken);
 
         private readonly ImmutableArray<T> _values;
         private readonly int _hashCode;
         private readonly int _length;
+
         public T this[int index] => _values[index];
         public int Count => _length;
         
@@ -137,7 +39,7 @@ namespace Datacute.IncrementalGeneratorExtensions
         public bool IsDefault => _values.IsDefault;
         public bool IsDefaultOrEmpty => _values.IsDefaultOrEmpty;
 
-        private EquatableImmutableArray(ImmutableArray<T> values, int hashCode)
+        internal EquatableImmutableArray(ImmutableArray<T> values, int hashCode)
         {
             _values = values;
             _length = values.Length;